Implement navbar switching with fragments instead of using activites
Switching navbar views using fragments allows the navbar to be persistent
instead of redrawn everytime a new activity is spawn. This redraw is jarring
from a UI perspective. In addition, modifying the code to switch activites in the
old BaseActivity was very janky since it needed to use the class names to start
new activities in the onNavigationItemSelected(). In addition, inflating a layout
when an activity is created is basically the same as inflating a frame layout for a
fragment so the old activity based implementation didn't make things simpler.