Add Ability to Take Pictures and Transform Them Between Format
Provides a controller to dispatch reqeuests to take an image from an
external camera app (in either an activity or a fragment) and then
read the result using a callback. Also, provides utilities for encoding
and decoding pictures in order to reduce boilerplate in the callback and
when sending data to Firebase.