Create EditProfileFragment
Creates a fragment that can be used as a form to update the user profile.
Updates the phone number and profile picture for the user.
Updates to email were left out since we use email to login and changing the
email in both Firebase Auth and Firebase RTDB is kind of a pain. I left it out
for now and will introduce it in a future PR. Since the specs didn't require the
ability to change the user name, I'm leaving it out.